minor fixes to MNL

+++ Demand/demand_utils/vax_entities.py
@@ -31,7 +31,7 @@ class Economy:
         self.abepsilon = [np.zeros(max_rank) for tt in range(n_geogs)] 
         # all-but-epsilon,  n_geogs x n_locs. abepsilon[tt][ll] = abd[tt] + distcoef * dists[tt][ll]. 
 
-        self.offers = [np.concatenate([[geog_pops[tt]], np.zeros(len(locs[tt])-1, dtype=int)]) for tt in range(n_geogs)]
+        self.offers = [np.concatenate([[geog_pops[tt]], np.zeros(len(locs[tt])-1)]) for tt in range(n_geogs)]
         # list of lists of location rankings offered, length n_geogs. offers[tt][ll] = number of individuals in tt offered location ranked ll
         # initialize with all individuals offered their nearest location for the first demand estimation
